FUTMAN-57
Moved around a lot of files, added a POM via Maven. Added test cases for each of the existing classes with no tests.
