+++ com/futman/src/main/.gitignore
@@ -1,0 +1,7 @@
+/Competition.class
+/Country.class
+/Match.class
+/Player$Position.class
+/Player.class
+/Season.class
+/Team.class
