+++ com/futman/src/main/Competition.java
@@ -1,3 +1,7 @@
+package com.futman.src.main;
+
+
+
 import java.util.ArrayList;
 
 public abstract class Competition {
