+++ com/futman/src/test/.gitignore
@@ -1,0 +1,6 @@
+/CompetitionTest.class
+/CountryTest.class
+/MatchTest.class
+/PlayerTest.class
+/SeasonTest.class
+/TeamTest.class
